requirements.txt
PyQt5
SQLAlchemy
python-dotenv
bcrypt
pymysql
pytest
